requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
torch>=1.9.0
faiss-cpu>=1.7.0
plotly>=5.0.0
